Babylonian square root - with fix

+++ examples/3B/fix_compile_error2/sqrt.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // This function computes the square root using the Babylonian method:
-float babylonian_sqrt (const double value, int& niter);
+double babylonian_sqrt (const double value, int& niter);
